+++ GEOSwift/GEOS/GEOSError.swift
@@ -3,4 +3,5 @@ public enum GEOSError: Error {
     case libraryError(errorMessages: [String])
     case wkbDataWasEmpty
     case typeMismatch(actual: GEOSObjectType?, expected: GEOSObjectType)
+    case noMinimumBoundingCircle
 }